+++ src/azure-cli-core/HISTORY.rst
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+
+2.0.52
+++++++
+* core: support cross tenant resource provisioning for multi-tenant service principal
+* Fix bug where ids piped from a command with tsv output is improperly parsed.
+
 2.0.51
 ++++++
 * msi login: do not reuse subscription name for identity info