+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.2.6
++++++
+* Minor fixes.
+
 2.2.5
 +++++
 * Improve handling of corner cases for storage copy commands.