+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ===============
 
+2.2.9
+++++++
+* `vm extension show / wait`: deprecated --expand parameter.
+* `vm restart`: Added `--force` which redeploys unresponsive VMs.
+* `vm/vmss create`: `--authentication-type` now accepts/infers "all" to create a VM with both password and ssh authentication.
+* `image create`: Added `--os-disk-caching` parameter to set os disk caching for an image.
+
 2.2.8
 ++++++
 * `vm/vmss create --storage-sku`: can now specify the storage account sku for managed os and data disks separately.